# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: TKA CBT Platform API
  version: 1.0.0
  license:
    name: Proprietary
    url: https://spdx.org/licenses/LicenseRef-Proprietary.html
  description: |
    Kontrak REST API untuk autentikasi, katalog dan pembayaran, CBT autosave,
    submit ujian, serta hasil analitik. Semua waktu menggunakan ISO 8601 UTC.
servers:
  - url: /
    description: Same host as the served API documentation
...
paths:
  /api/v1/auth/register:
    post:
      tags: [Authentication]
      summary: Register student account
      operationId: registerUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Account successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/auth/login:
    post:
      tags: [Authentication]
      summary: Login and establish the active device session
      description: A successful login replaces the previous JTI stored in Redis.
      operationId: loginUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/auth/logout:
    post:
      tags: [Authentication]
      summary: Revoke the active session
      operationId: logoutUser
      security:
        - BearerAuth: []
      responses:
        '204':
          description: Session successfully revoked
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/packages:
    get:
      tags: [Packages]
      summary: List available tryout packages
      operationId: listPackages
      security: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Package list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PackageListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/transactions/checkout:
    post:
      tags: [Commerce]
      summary: Create a package checkout transaction
      operationId: checkoutPackage
      security:
        - BearerAuth: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          description: Unique key generated by the client to prevent duplicate checkout.
          schema:
            type: string
            minLength: 16
            maxLength: 128
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckoutRequest'
      responses:
        '201':
          description: Transaction created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckoutResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/webhooks/payment:
    post:
      tags: [Commerce]
      summary: Receive a payment gateway status notification
      description: |
        The provider adapter verifies the callback (Xendit `x-callback-token`,
        plus optional `X-Payment-Timestamp`/`X-Payment-Signature`) and translates
        the provider payload into the canonical `PaymentWebhookRequest`. Events
        are processed idempotently by `event_id`.
      operationId: paymentWebhook
      security:
        - WebhookToken: []
      parameters:
        - name: X-Callback-Token
          in: header
          required: true
          description: Webhook callback token configured for the payment provider.
          schema:
            type: string
        - name: X-Payment-Timestamp
          in: header
          required: false
          description: Unix timestamp in seconds; when supplied, requests older than five minutes are rejected.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentWebhookRequest'
      responses:
        '200':
          description: Webhook accepted, including previously processed events
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Invalid webhook signature
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/admin/payment-settings:
    get:
      tags: [Administration]
      summary: Get the active payment-provider configuration
      description: Owner-only. Secret values are never returned in plaintext.
      operationId: getPaymentSettings
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Masked payment-provider configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentSettingsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags: [Administration]
      summary: Update the active payment-provider configuration
      description: |
        Owner-only. Omitted credentials and endpoint fields retain their
        current values. An explicitly empty redirect URL removes that redirect.
        Saved credentials are encrypted and activated without restarting the API.
      operationId: updatePaymentSettings
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentSettingsInput'
      responses:
        '200':
          description: Configuration saved and activated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentSettingsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/exams/{id}/start:
    get:
      tags: [CBT]
      summary: Start or resume an exam
      description: |
        The path id is `exams.id`. This operation is idempotent while an ongoing
        attempt exists and returns time derived from the server clock.
      operationId: startExam
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Exam UUID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Exam attempt started or resumed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExamStartResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/cbt/answers/sync:
    post:
      tags: [CBT]
      summary: Synchronize one or more answers
      description: Idempotent upsert used by the real-time autosave client.
      operationId: syncAnswers
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnswerSyncRequest'
      responses:
        '200':
          description: Answers synchronized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnswerSyncResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/exams/{id}/submit:
    post:
      tags: [CBT]
      summary: Submit an exam attempt
      description: The path id is `user_exams.id`, not the exam master ID.
      operationId: submitExam
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: User exam attempt UUID
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitExamRequest'
      responses:
        '200':
          description: Attempt submitted and scored
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmitExamResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/exams/{id}/result:
    get:
      tags: [Analytics]
      summary: Get a submitted exam result
      description: The path id is `user_exams.id`.
      operationId: getExamResult
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: User exam attempt UUID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Exam result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExamResultResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Exam is still ongoing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/rankings/global:
    get:
      tags: [Analytics]
      summary: Get the global student ranking
      description: Returns each student's best submitted exam score, ordered from highest to lowest.
      operationId: getGlobalRanking
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Global ranking, limited to 100 students
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/GlobalRankingEntry'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token. The token JTI must match the active Redis session.
    WebhookSignature:
      type: apiKey
      in: header
      name: X-Payment-Signature
      description: HMAC signature supplied by the payment provider (optional).
    WebhookToken:
      type: apiKey
      in: header
      name: X-Callback-Token
      description: Webhook callback token that must match PAYMENT_WEBHOOK_SECRET.
  responses:
    BadRequest:
      description: Malformed request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing, invalid, expired, or replaced session token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: SESSION_INVALID
            message: Device Limit Reached / Session Invalid
            request_id: req_01JXYZ
    Forbidden:
      description: Authenticated user cannot access this resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Request conflicts with the current resource state
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    TooManyRequests:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds until another request may be attempted
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    RegisterRequest:
      type: object
      additionalProperties: false
      required: [email, password, school_level]
      properties:
        email:
          type: string
          format: email
          maxLength: 320
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 72
          writeOnly: true
        school_level:
          type: string
          enum: [SD, SMP, SMA]
    RegisterResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: '#/components/schemas/User'
    LoginRequest:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          writeOnly: true
        device_id:
          type: string
          description: Optional client identifier for audit metadata; JTI remains authoritative.
          maxLength: 200
    LoginResponse:
      type: object
      required: [access_token, token_type, expires_in, user]
      properties:
        access_token:
          type: string
        token_type:
          type: string
          enum: [Bearer]
        expires_in:
          type: integer
          format: int64
          description: Access token lifetime in seconds
        user:
          $ref: '#/components/schemas/User'
...
    PackageListResponse:
      type: object
      required: [data, page, per_page]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Package'
        page:
          type: integer
        per_page:
          type: integer
    CheckoutRequest:
      type: object
      additionalProperties: false
      required: [package_id, payment_method]
      properties:
        package_id:
          type: string
          format: uuid
        payment_method:
          type: string
          enum: [qris, virtual_account, e_wallet]
...
    CheckoutResponse:
      type: object
      required: [transaction, payment_url, expires_at]
      properties:
        transaction:
          $ref: '#/components/schemas/Transaction'
        payment_url:
          type: string
          format: uri
        qr_string:
          type: string
          nullable: true
        expires_at:
          type: string
          format: date-time
    PaymentWebhookRequest:
      type: object
      additionalProperties: false
      required: [event_id, invoice_number, payment_status, amount]
      properties:
        event_id:
          type: string
          maxLength: 200
        invoice_number:
          type: string
          maxLength: 100
        payment_status:
          type: string
          enum: [paid, failed, expired, refunded]
        payment_method:
          type: string
          nullable: true
        amount:
          type: number
          format: double
          minimum: 0
        paid_at:
          type: string
          format: date-time
          nullable: true
    WebhookResponse:
      type: object
      required: [received, processed]
      properties:
        received:
          type: boolean
        processed:
          type: boolean
          description: False when event_id was already processed.
...
    PaymentSettingsResponse:
      type: object
      required: [settings]
      properties:
        settings:
          $ref: '#/components/schemas/PaymentSettings'
...